Improve database interaction
- I am still figuring out a way to pass the database around.
- The next commit will involve the reader monad. :)
