+++ subprojects/webapp/src/main/scala/at/bioinform/webapp/Application.scala
@@ -1,25 +1,25 @@
 package at.bioinform.webapp
 
-import slick.jdbc.H2Profile.api._
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.marshalling.Marshal
-import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpResponse}
+import akka.http.scaladsl.model.{ContentTypes, HttpEntity}
 import akka.http.scaladsl.server.Directives._
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.Flow
 import akka.util.ByteString
 import at.bioinform.lucene.segment.Segment
-import at.bioinform.model.LuceneIndex
+import at.bioinform.model.db.TableDefinitions
 import at.bioinform.stream.fasta.{FastaFlow, FastaSegmenter}
+import at.bioinform.webapp.directives.LuceneIndexResource
 import com.typesafe.config.ConfigFactory
 import org.slf4j.LoggerFactory
+import slick.jdbc.H2Profile.api._
 
-import scala.concurrent.Await
 import scala.concurrent.duration.Duration
+import scala.concurrent.{Await, Future}
 import scala.io.StdIn
 
-object Application {
+object Application extends TableDefinitions with LuceneIndexResource {
 
   val Logger = LoggerFactory.getLogger(Application.getClass.getName)
 
@@ -54,29 +54,9 @@ object Application {
               complete(HttpEntity(ContentTypes.`text/html(UTF-8)`, source))
             }
           }
-        } ~
-    path("index" ) {
-      post {
-        complete {
-          import akka.http.scaladsl.marshalling.GenericMarshallers._
-          val db = Database.forConfig("h2mem1")
-          try {
-            val insertQuery = TableQuery[LuceneIndex] returning TableQuery[LuceneIndex].map(e => e.id) into ((index, id) => index.copy(_1 = id))
-            val insert = insertQuery += (0, "first", "path1")
-            Marshal(db.run(insert).map(_.toString())).to[HttpResponse]
-          } finally db.close
-        }
-      }
-    }
-    // Use H2Profile to connect to an H2 database
-    val setup = DBIO.seq(
-      TableQuery[LuceneIndex].schema.create,
-    )
+        } ~ luceneIndexRoutes
 
-    Logger.info("Creating database schema")
-    val db = Database.forConfig("h2mem1")
-    val setupFuture = db.run(setup)
-    Await.result(setupFuture, Duration.Inf)
+    Await.result(createDbSchema(), Duration.Inf)
 
     Logger.info(s"Server online at http://localhost:8080/\nPress RETURN to stop...")
     val bindingFuture = Http().bindAndHandle(route, "localhost", 8080)
@@ -87,4 +67,9 @@ object Application {
       .onComplete(_ => system.terminate()) // and shutdown when done
   }
 
+  def createDbSchema(): Future[Unit] = {
+    Logger.info("Creating database schema")
+    val db = Database.forConfig("h2mem1")
+    db.run(schema)
+  }
 }
